Bug Fix on Usage of HistoryManager
- Fixed bug that caused the active symbol to lose interactivity
unexpectedly upon redoing an 'add' or undoing a 'remove'. This happened
because after inserting the subtree of symbols (previously removed by
undoing an 'add' or performing a 'remove'), the root of the subtree
would not be selected as the active element, the group containing it
would (which was wrong by design).