Added Support for Undo/Redoing Fine Symbol Reshape + Bug Fix
- Application can now undo/redo all fine symbol reshape types available
in layer controller (side stretch, diagonal stretch, and side shear).
- Fixed bug related to passing arrays by reference that caused history
pushed into history manager not to be final and thus, be wrongfully
modified in later use of the application. More specifically, undoing and
redoing would cause "symbol_reshape" saved actions to replace layer data
(symbol vertices array)  with a reference to its own array saved in
memory, instead of copying the values over to the layer.