Rename Function Slight Redesign + Bug Fix
- Changed use of rename function in Layer Manager. Previously every
layer/group had a callback to it and context menu called that from the
element, now context menu calls Layer Manager sending the element to be
renamed.
- Fixed bug that caused UI not to be updated when renaming the name of
the symbol art. This was due to the function called to initialize the
main group of the symbol art not enveloping the name of the group inside
a <span> tag like every other element created.