Added Cross Application Symbol Alpha Compatibility
- Application now supports loading from and saving to any of the two
supported applications (this one and the third-party .NET Symbol Art
Editor application described in README.md) WITH layer transparency !
Previously, this was an issue for the third-party application due to a
bug it has: it can load and save .sar WITH transparency, but it cannot
comprehend layer transparency from .saml files it itself created. This
application now corrects that issue (thanks to reverse engineering) and
now supports both the buggy .saml format from the third
party-application AND the corrected .saml format created by this
application.