Improved Toolbar + UI Improvement + Bug Fix
- Toolbar now supports any number of tool the developer decides it to
have. It automatically finds the best fit for the current screen
resolution and if that best fit cannot fit all tools, it enables the
swipe feature which lets users reach every tool by swiping the toolbar
left and right.
- UI Improvement: Select menu now has 85% opacity on big viewports and
70% opacity on small viewports to allow users to still see the symbol
art even when it is on top of it.
- UI Improvement: Layer controller and color picker are now hidden
whenever selectmenu is open and on a small viewport. This was done to
avoid them to go off screen or take unexpected positions on viewport.
- Bug Fix: Layer controller is now controlled by transform css instead
of float due to it taking unexpected position on screen when there is
not enough width for it. It was necessary because this impacted other
features such as the position of the layer edit box.