Added Layer Capacity + Layer Count Display + Bug Fix
- Now the application allows up to 225 symbols to be added per symbol
art. (as per specs of original symbol art editor).
- Layer Manager now uses its header to display the current count of
layers in editor.
- Layer Manager header now sticks to the top of screen even when
scrolling to the bottom of the list so it is always visible.
- Fixed bug that caused uploaded files with unsupported extensions to
result in an runtime error. It is now caught, logged, and alerted to
user that loading failed.
- Fixed bug that would disable normal logging if .saml upload failed.
This happened when the upload failed after temporarily disabling logging
but before enabling it again.