Added Navigation Indicators on Toolbar
- Toolbar now displays little circles (filled or hollowed) on top of it
to indicate which "page" of if it is on. This only shows when toolbar
gets resized to fit a screen smaller than desired and there is a need
for swiping left/right to use all tools. Navigation now helps see how
many pages there are and which it is currently in.