Added Copy/Paste Feature + Added Alert Manager + Bug Fix
- Application now supports copying/pasting layers and groups by using
the newly added copy/paste options in context menu (only the symbol art
folder itself cannot be copied).
- Application supports undo/redoing paste actions.
- Added AlertManager to manage friendly in-app alerts. It works by
pushing pending alerts which pop up in sequence in the app until no
pending alert is left. It currently has a capacity to 40 alerts, any
further push is ignored until subsequent alert pops. The time each alert
remains in screen depends on the average reading speed and the length of
the message.
- Reorganized context menu options, renamed a few options, changed a few
hotkeys, and added the new copy/paste options.
- Fixed bug on history management symbol_groupmove action that led it to
sometimes reference outdated information in editor. It used a reference
to a editor.layers, which can get overwritten, thus outdating the
reference.
- Added extra param to layer and group init for purposes of copy/paste.
- Added pastefrom function in layer for purposes of copy/paste.
- Fixed bug that caused error on mobile when trying to make a ghost
image of layer/groups to move (which in this case, it should not in the
first place).
- Fixed bug that caused normal logging to be disabled on load of .saml.