Added Overlay Image Feature + Bug Fix
- Added overlay image feature:
- Users can now upload an image to the application and edit the work
with the assistance of a helper image leading their way to success.
- Added dedicated toolbar options for uploading an overlay image and
toggling the overlay controller.
- Added dedicated controller that allows users to change transparency,
size, and rotation of the uploaded overlay image, and even add/remove a
green screen as background to help visualize the work. While controller
is active, all other controllers are deactivated. The reverse also
applies.
- While overlay controller is active, users can move the uploaded
overlay image by dragging the image on editor.
- Overlay used is also saved in .saml file for ease when resuming the
work. This means .saml files can now contain overlay image information,
which will be loaded info the application in the exact position,
rotation, size, transparency used when it was saved.
- Added enable/disable for tool options in toolbar (80% accuracy due to
limitation in how jQuery toolbar plugin used was implemented).
- Fixed bug that caused accelerated and retarded motion of groups of
symbols in mobile devices. The issue was due to editor zoom not being
taken into account.