Introducing Group Edit Box + Symbol Usage Recall
- Added corner buttons for group selection (purely cosmetic yet) as
visual cues of where the corners of the group container are located in
the editor. These buttons may later be used for editting groups of
symbols at once.
- When adding new symbols, the application now uses the most recently
symbol shape and color picked. So if the user recently reshaped a symbol
to a black square and then added a new symbol, the new symbol will show
up as a black square as default instead of the previous white circle.